+++ code/renderer/tr_public.h
@@ -151,7 +151,7 @@ typedef struct {
 															const int passEntityNum, const int contentmask, 
 															const EG2_Collision eG2TraceType, const int useLod );
 
-	qboolean			(*SG_Append)						( unsigned long chid, const void *pvData, int iLength );
+	qboolean			(*SG_Append)						( unsigned int chid, const void *pvData, int iLength );
 
 	int					(*SV_PointContents)					( const vec3_t p, clipHandle_t model );
